+++ src/main/java/com/mobble/mobbleserver/domain/notification/outbox/Status.java
@@ -1,0 +1,5 @@
+package com.mobble.mobbleserver.domain.notification.outbox;
+
+public enum Status {
+    PENDING, SENT, FAILED
+}
