--- src/main/java/com/mobble/mobbleserver/refactor/notification/outbox/entity/Status.java
@@ -1,5 +1,0 @@
-package com.mobble.mobbleserver.refactor.notification.outbox.entity;
-
-public enum Status {
-    PENDING, SENT, FAILED
-}
